requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.2
PyMySQL==1.1.0
cryptography==41.0.7
bcrypt==4.1.1
python-dotenv==1.0.0
gunicorn==21.2.0
